all but last line working properly
