+++ browser/app.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Routes from './react/Routes';
+import Routes from './react-redux/Routes';
 
 import { Provider } from 'react-redux';
-import reducer from './react/reducers';
+import reducer from './react-redux/reducers';
 import { createStore, applyMiddleware } from 'redux';
 import createLogger from 'redux-logger';
 import thunkMiddleware from 'redux-thunk';
