+++ browser/react-redux/actions/characters.js
@@ -1,1 +1,26 @@
-characters.js
+/* ========== CONSTANTS ========== */
+const LOAD_CHARACTERS = 'LOAD_CHARACTERS';
+const SET_CURRENT_CHARACTER = 'SET_CURRENT_CHARACTER';
+
+/* ========== ACTION CREATORS ========== */
+export const loadCharacters = characters => ({
+	type: LOAD_CHARACTERS,
+	characters
+});
+
+export const setCurrentCharacter = character => ({
+	type: SET_CURRENT_CHARACTER,
+	character
+})
+
+/* ========== ASYNC ========== */
+export const fetchCharacters = playName => {
+	return dispatch => {
+		fetch(`/api/plays/${playName}/characters`)
+		.then(res => res.json())
+		.then(characters => {
+			dispatch(loadCharacters(characters))
+		})
+		.catch(err => console.error(err.stack));
+	}
+}
