+++ browser/react-redux/actions/lines.js
@@ -1,1 +1,8 @@
-lines.js
+/* ========== CONSTANTS ========== */
+const SET_CURRENT_LINE = 'SET_CURRENT_LINE';
+
+/* ========== ACTION CREATORS ========== */
+export const setCurrentLine = line => ({
+	type: SET_CURRENT_LINE,
+	line
+});
