+++ browser/react-redux/containers/LineContainer.js
@@ -1,6 +1,7 @@
 import Line from '../components/Line';
 import { connect } from 'react-redux';
-import { startPlayingFromLine, listenToLine, stopSpeakingLine } from '../actions';
+import { startSpeakingFromLine,stopSpeakingLine } from '../actions/speaking';
+import { listenToLine } from '../actions/listening';
 
 const mapStateToProps = ({ currentPlay, currentLine, isListening, isSpeaking }, { line }) => {
 	return {
@@ -14,7 +15,7 @@ const mapStateToProps = ({ currentPlay, currentLine, isListening, isSpeaking }, 
 
 const mapDispatchToProps = dispatch => ({
 	toggleLine(line, play, isSpeaking) {
-		if (!isSpeaking) dispatch(startPlayingFromLine(line, play));
+		if (!isSpeaking) dispatch(startSpeakingFromLine(line, play));
 		else dispatch(stopSpeakingLine());
 	},
